requirements.txt
aiopg==1.3.1
alembic==1.6.2
apispec==4.4.1
appdirs==1.4.4
asgiref==3.4.1
async-timeout==3.0.1
attrs==21.2.0
Babel==2.9.1
backports.entry-points-selectable==1.1.1
bcrypt==3.2.0
beautifulsoup4==4.9.3
bidict==0.21.4
black==21.5b1
blinker==1.4
boto3==1.23.4
botocore==1.26.4
bs4==0.0.1
certifi==2020.12.5
cffi==1.14.5
cfgv==3.3.0
chardet==4.0.0
click==8.0.1
croniter==1.0.15
cryptography==3.4.7
distlib==0.3.1
Django==3.2.5
et-xmlfile==1.1.0
filelock==3.4.2
flake8==3.9.2
Flask==2.0.0
Flask-Babel==2.0.0
Flask-Caching==1.10.1
Flask-Cors==3.0.10
Flask-JWT-Extended==4.2.1
flask-marshmallow==0.14.0
Flask-Migrate==3.0.0
flask-smorest==0.31.0
Flask-SocketIO==5.1.1
Flask-SQLAlchemy==2.5.1
greenlet==1.1.0
h11==0.12.0
identify==2.2.4
idna==2.10
importlib-metadata==4.12.0
iniconfig==1.1.1
install==1.3.5
itsdangerous==2.0.1
Jinja2==3.0.1
jmespath==1.0.0
keyring==23.7.0
Mako==1.1.4
MarkupSafe==2.0.1
marshmallow==3.12.1
marshmallow-enum==1.5.1
marshmallow-sqlalchemy==0.26.1
marshmallow-union==0.1.15.post1
mccabe==0.6.1
mypy==0.812
mypy-extensions==0.4.3
nodeenv==1.6.0
oauthlib==3.1.1
openpyxl==3.0.9
packaging==20.9
paramiko==2.11.0
pathspec==0.8.1
Pillow==8.3.1
platformdirs==2.4.1
pluggy==0.13.1
pre-commit==2.12.1
procrastinate==0.20.0
psycopg2-binary==2.9.3
py==1.10.0
pycodestyle==2.7.0
pycparser==2.20
pyflakes==2.3.1
PyJWT==2.1.0
PyNaCl==1.5.0
pyparsing==2.4.7
pytest==6.2.4
python-dateutil==2.8.1
python-dotenv==0.17.1
python-editor==1.0.4
python-engineio==4.3.0
python-magic-bin==0.4.14
python-socketio==5.5.0
pytz==2021.1
PyYAML==5.4.1
redis==3.5.3
regex==2021.4.4
requests==2.25.1
requests-oauthlib==1.3.0
s3transfer==0.5.2
sentry-sdk==1.3.1
simple-websocket==0.5.0
six==1.16.0
soupsieve==2.2.1
SQLAlchemy==1.4.15
sqlparse==0.4.1
toml==0.10.2
tomli==2.0.0
typed-ast==1.4.3
typing-extensions==3.10.0.0
tzlocal==2.1
urllib3==1.26.4
virtualenv==20.10.0
webargs==8.0.0
Werkzeug==2.0.1
wsproto==1.0.0
zipp==3.8.1
